+++ src/app/routes/transactions/placeOrder.ts
@@ -563,7 +563,7 @@ export async function authorizePointAward(req: Request) {
         const givePointAwardParams: cinerino.factory.transaction.placeOrder.IGivePointAwardParams[] = [];
 
         for (const programMembership of programMemberships) {
-            const membershipServiceId = <string>programMembership.membershipFor?.id;
+            const membershipServiceId = <string>(<any>programMembership).membershipFor?.id;
             const membershipService = await productService.findById({ id: membershipServiceId });
 
             // 登録時の獲得ポイント
@@ -583,7 +583,7 @@ export async function authorizePointAward(req: Request) {
                             sort: { ownedFrom: cinerino.factory.sortType.Ascending },
                             limit: 1,
                             typeOfGood: {
-                                typeOf: cinerino.factory.ownershipInfo.AccountGoodType.Account,
+                                typeOf: cinerino.factory.chevre.paymentMethodType.Account,
                                 accountType: membershipPointsEarnedUnitText
                             },
                             ownedBy: { id: req.agent.id },
@@ -596,19 +596,21 @@ export async function authorizePointAward(req: Request) {
                     });
 
                     // 開設口座に絞る
-                    accountOwnershipInfos = accountOwnershipInfos.filter(
-                        (o) => o.typeOfGood.status === cinerino.factory.pecorino.accountStatusType.Opened
-                    );
+                    accountOwnershipInfos = accountOwnershipInfos.filter((o) => {
+                        return (<cinerino.factory.pecorino.account.IAccount>o.typeOfGood).status
+                            === cinerino.factory.pecorino.accountStatusType.Opened;
+                    });
                     if (accountOwnershipInfos.length === 0) {
                         throw new cinerino.factory.errors.NotFound('accountOwnershipInfos');
                     }
-                    const toAccount = accountOwnershipInfos[0].typeOfGood;
+                    const toAccount = <cinerino.factory.pecorino.account.IAccount>accountOwnershipInfos[0].typeOfGood;
 
                     givePointAwardParams.push({
                         object: {
                             typeOf: cinerino.factory.action.authorize.award.point.ObjectType.PointAward,
                             amount: membershipPointsEarnedValue,
                             toLocation: {
+                                typeOf: cinerino.factory.chevre.paymentMethodType.Account,
                                 accountType: membershipPointsEarnedUnitText,
                                 accountNumber: toAccount.accountNumber
                             },