+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - SendGrid設定に関して、プロセスレベルでの設定とプロジェクトレベルでの設定を両方有効化
+- 所有権インターフェースの汎用性拡張
+- Chevre転送取引へのignorePaymentCard指定を削除
 
 ### Deprecated